+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Unreleased
 
 - Added `background_rule` to the parameter of `draw_keypoint`.
+- Added exception handling for getting token.
 
 ## 1.1.0